Jakarta Data Integration - Restructure classes in JNoSQL Integration
JPA classes in a simpler package name, NoSQL classes in a separate package
