+++ appserver/persistence/jnosql-integration/src/main/java/org/glassfish/main/jnosql/jakartapersistence/EntityValidatorInterceptor.java
@@ -13,7 +13,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package org.glassfish.main.jnosql.jakartapersistence.mapping.glassfishcontext;
+package org.glassfish.main.jnosql.jakartapersistence;
 
 import jakarta.interceptor.InvocationContext;
 import jakarta.validation.ConstraintViolation;
