+++ appserver/persistence/jnosql-integration/src/main/resources/META-INF/services/org.eclipse.jnosql.jakartapersistence.mapping.metadata.JakartaPersistenceClassScanner
@@ -1,1 +1,1 @@
-org.glassfish.main.jnosql.jakartapersistence.mapping.glassfishcontext.GlassFishJakartaPersistenceClassScanner
+org.glassfish.main.jnosql.jakartapersistence.GlassFishJakartaPersistenceClassScanner
