+++ appserver/persistence/jnosql-integration/src/main/resources/META-INF/services/org.eclipse.jnosql.mapping.metadata.ClassScanner
@@ -1,1 +1,1 @@
-org.glassfish.main.jnosql.jakartapersistence.mapping.glassfishcontext.GlassFishNosqlClassScanner
+org.glassfish.main.jnosql.nosql.GlassFishNosqlClassScanner
