+++ appserver/persistence/jnosql-integration/src/test/java/org/glassfish/main/jnosql/jakartapersistence/MyOtherEntity.java
@@ -13,12 +13,18 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package org.glassfish.main.jnosql.jakartapersistence.mapping.glassfishcontext;
+package org.glassfish.main.jnosql.jakartapersistence;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 
 /**
  *
  * @author Ondro Mihalyi
  */
-public class MyUnsupportedEntity {
+@Entity
+public class MyOtherEntity {
 
+    @Id
+    long id;
 }
