+++ appserver/persistence/jnosql-integration/src/test/java/org/glassfish/main/jnosql/jakartapersistence/repositories/MyNoInterfaceRepository.java
@@ -13,18 +13,15 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
-package org.glassfish.main.jnosql.jakartapersistence.mapping.glassfishcontext.repositories;
+package org.glassfish.main.jnosql.jakartapersistence.repositories;
 
-import jakarta.data.repository.CrudRepository;
 import jakarta.data.repository.Repository;
-
-import org.glassfish.main.jnosql.jakartapersistence.mapping.glassfishcontext.MyEntity;
 
 /**
  *
  * @author Ondro Mihalyi
  */
 @Repository
-public interface MyCrudRepository extends CrudRepository<MyEntity, Long>{
+public interface MyNoInterfaceRepository {
 
 }
